requirements.txt
Appium-Python-Client~=2.2.0
pytest~=7.1.2
selenium~=4.1.5
allure-pytest
slackweb~=1.0.5
